### 2021-11-02 #3858 @leouieda
👋🏽 Hi @valentineap thank you for your submission to JOSS! I'll first do some light editorial proof-reading of your paper and package and then search for suitable reviewers. In the mean time, you can:

1. Suggest suitable reviewers (who you already know or from [our list of reviewers](https://bit.ly/joss-reviewers)). Please include their GitHub username without @-mentioning to avoid spamming their notifications (e.g. `leouieda` instead of @leouieda).
2. Please include the missing DOIs in your references picked up by our bot Whedon above.
3. Please also add citations to numpy and scipy, which you mention as dependencies of pyprop8.

When you push changes to your paper, you can ask Whedon to regenerate the PDF by commenting `@whedon generate pdf` here. You can also use the preview service to test paper compilation in a separate branch before merging: https://whedon.theoj.org/
