+++ paper.md
@@ -21,7 +21,7 @@ date: 9 October 2021
 bibliography: paper.bib
 ---
 # Summary
-The package `pyprop8` enables calculation of the response of a 1-D layered halfspace to a seismic source, and also derivatives ('sensitivity kernels') of the wavefield with respect to source parameters. Seismograms, seismic spectra, and measures of static displacement (e.g. GPS, InSAR and field observations) may all be simulated. The method is based on a Thompson-Haskell propagator matrix algorithm, described in @OToole2011 and @OToole2012. The package is entirely written in Python, dependent only on the mainstream libraries `numpy` and `scipy`. As such it is lightweight and easy to deploy across a variety of platforms, making it particularly suited to use for teaching and outreach purposes.
+The package `pyprop8` enables calculation of the response of a 1-D layered halfspace to a seismic source, and also derivatives ('sensitivity kernels') of the wavefield with respect to source parameters. Seismograms, seismic spectra, and measures of static displacement (e.g. GPS, InSAR and field observations) may all be simulated. The method is based on a Thompson-Haskell propagator matrix algorithm, described in @OToole2011 and @OToole2012. The package is entirely written in Python, dependent only on the mainstream libraries `numpy` [@Harris2020] and `scipy` [@Virtanen2020]. As such it is lightweight and easy to deploy across a variety of platforms, making it particularly suited to use for teaching and outreach purposes.
 
 # Statement of need
 Many different tools and packages exist which may be used to simulate seismic wave propagation. Generally, these are designed to support the needs of the seismological research community. As such, they tend to be optimised for computational efficiency, and are partly or entirely dependent on code written in compiled languages. In many cases they interface with a variety of external libraries---for example, those designed to implement particular data formats. As such, installation, deployment and use can be cumbersome, especially for non-expert users and on non-Unix-like operating systems. This can create substantial barriers in contexts such as teaching and outreach, where it may be necessary to accommodate and support users across a wide variety of platforms.
