+++ paper.bib
@@ -5,7 +5,8 @@
 	pages = {1516--1536},
 	title = {Numerically stable computation of complete synthetic seismograms including the static displacement in plane layered media},
 	volume = {187},
-	year = {2011}
+	year = {2011},
+	doi = {10.1111/j.1365-246X.2011.05210.x}
 }
 
 @article{OToole2012,
@@ -14,7 +15,8 @@
 	pages = {257--270},
 	title = {Centroid--Moment Tensor Inversions Using High-Rate {GPS} Waveforms},
 	volume = {191},
-	year = {2012}
+	year = {2012},
+	doi = {10.1111/j.1365-246X.2012.05608.x}
 }
 
 @article{OToole2013,
@@ -23,7 +25,8 @@
 	pages = {60--65},
 	title = {Earthquake source parameters from {GPS}-measured static displacements with potential for real‐time application},
 	volume = {40},
-	year = {2013}
+	year = {2013},
+	doi = {10.1029/2012GL054209}
 }
 
 @article{Kaufl2014,
@@ -32,7 +35,8 @@
 	pages = {1676--1693},
 	title = {A framework for fast probabilistic centroid--moment-tensor determination --- Inversion of regional static displacement measurements},
 	volume = {196},
-	year = {2014}
+	year = {2014},
+	doi = {10.1093/gji/ggt473}
 }
 
 @inproceedings{OToole2013a,
@@ -40,7 +44,8 @@
   title = {Microseismic Source Mechanisms: What the Waveform Can Tell Us},
   publisher = {European Association of Geoscientists and Engineers},
   year = {2013},
-  booktitle = {Proceedings of the  4th EAGE Passive Seismic Workshop}
+  booktitle = {Proceedings of the  4th EAGE Passive Seismic Workshop},
+	doi = {10.3997/2214-4609.20142350}
 }
 
 
@@ -50,7 +55,8 @@
 	pages = {2299--2312},
 	title = {Robust and fast probabilistic source parameter estimation from near-field displacement waveforms using pattern recognition},
 	volume = {105},
-	year = {2015}
+	year = {2015},
+	doi = {10.1785/0120150010}
 }
 
 
@@ -60,7 +66,8 @@
 	pages = {1710--1728},
 	title = {Solving probabilistic inverse problems rapidly with prior samples},
 	volume = {205},
-	year = {2016}
+	year = {2016},
+	doi = {10.1093/gji/ggw108}
 }
 
 @incollection{Woodhouse1980a,
@@ -69,3 +76,46 @@
 	publisher = {Societ{\`a} Italiana di Fisica},
 	title = {Efficient and stable methods for performing seismic calculations in stratified media},
 	year = {1980}}
+
+
+@Article{Harris2020,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ journal       = {Nature},
[…]
