### 2021-11-03 #3858 @valentineap
Thanks for taking this on, @leouieda!

DOIs now fixed and scipy/numpy references added.

Reviewers: I'm a bit stuck here - a couple of names come to mind, but tracking those people down on github is proving difficult! I'm sure many of the people on the list of reviewers tagged 'seismology' (or perhaps even 'geodesy') would be suitable, but I don't recognise many of the names.
