### 2021-12-16 #3858 @leouieda
:wave: @seisman @sgkang would you be interested in reviewing this submission for [JOSS](https://joss.theoj.org/)? This paper is about a software package that does forward modelling and calculates gradients (for inversion) of seismic waves in 1D layered media. It’s meant as a research and teaching/outreach tool. Your combined expertise in seismology, inverse problems, and education would be much appreciated 🙂

Reviews are done entirely through GitHub (issues and pull requests). Please refer to our reviewer guidelines for more information about JOSS and our review process (feel free to ask me any questions as well): https://joss.readthedocs.io/en/latest/review_criteria.html Reviews usually take 2-4 weeks but we can certainly make accommodations if more time is needed (particularly this close to the end of the year holidays).

If you are not able to review at this time, please let me know if you have recommendations for other reviewers. **Please recommend without mentioning a handle to avoid notifying the person** (e.g., you would refer to me as leouieda instead of @leouieda). This way we attempt to reduce the number of people who are subscribed to this thread.
