### 2021-12-17 #3858 @seisman
@leouieda Hi Leo, thanks for your invitation. It looks really interesting, but I'm afraid I don't have much time in the next few months.

I would like to recommend my former colleague, Dr. Zhou Lu (GitHub handle: `luzhou-git`, Google Scholar: https://scholar.google.com/citations?user=V2Ijx9EAAAAJ), for review. He is a seismologist working on stress and crustal deformations, and also has a lot of experience in simulating earthquake dynamic and static deformations. I believe he is qualified for reviewing this submission.
