### 2021-12-20 #3858 @leouieda
@seisman thanks for the suggestion and quick response 🙂

:wave: @luzhou-git would you be interested in reviewing this submission for JOSS? Please see my comment above for a summary of the paper and instructions on how JOSS reviews work. Your expertise in deformation modelling would be much appreciated.
