### 2022-03-03 #3858 @hemmelig
Hi @leouieda - sounds good, I'm happy to take on this review. Is there any additional information beyond that available on the JOSS review criteria page of which I should be aware, as a first time reviewer? For example, is there an official review template? I have read the review checklist page of the JOSS documentation and was planning to just base my review around that. Thanks.
