### 2022-03-03 #3858 @leouieda
👋🏽  @hfmark would you be interested in reviewing this submission for JOSS? Please see [my comment above](https://github.com/openjournals/joss-reviews/issues/3858#issuecomment-995500979) for a summary of the paper and instructions on how JOSS reviews work. Your expertise in seismology and experience with Python software would be appreciated!
