### 2022-03-03 #3858 @leouieda
@meghanrjones thank you for considering and for the recommendation!

@hemmelig thank you agreeing to review! The JOSS review is a bit different from traditional reviews. There will be a checklist with criteria for you to fill out (that comes after reviewers are assigned and we actually start the review process). Reviewers are also encouraged to open issues and pull requests in the source repository to work with authors towards meeting all of the JOSS requirements and improving the software overall. So there is no template for a one-off review and this is more of a conversation between authors, reviewers, and editors. Which is kind of how we work on open-source in general (and on purpose).

For now, we just need to see if we can get a second reviewer assigned before starting the review. If you have any questions about the process at any time, I’m more than happy to answer 🙂
