### 2022-03-03 #3858 @hfmark
Hi @leouieda, I'm happy to review this. I also haven't reviewed for JOSS before - I'll look over the documentation.
