### 2022-03-04 #3858 @leouieda
Thank you @hfmark! 🚀  Please see [our reviewer guide](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more details and feel to ask any time.  I'll post a comment with more general instructions as soon as I start the review thread.
