### 2022-03-04 #4217 @editorialbot
[REVIEW]: pyprop8: A lightweight code to simulate seismic observables in a layered half-space
**Submitting author:** @valentineap (<a href="http://orcid.org/0000-0001-6134-9351">Andrew Valentine</a>)
**Repository:** https://github.com/valentineap/pyprop8
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.2
