### 2022-03-04 #4217 @leouieda
:wave: @valentineap @hemmelig @hfmark welcome to the review thread for the paper! All of our communications will happen here from now on.

**Before we start:** Both reviewers need to post the comment `@editorialbot generate my checklist` in this thread to generate their review checklist.

The JOSS review is different from most other journals:

* Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission.
* As such, the reviewers are **encouraged to submit issues and pull requests** on the software repository.
* When doing so, please **mention** `openjournals/joss-reviews#REVIEW_NUMBER` so that a link is created to this thread (and I can keep an eye on what is happening).
* As you go over the submission, please check any items in your list that you feel have been satisfied.
* Feel free to comment and ask questions on this thread.
* In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.
* We aim for reviews to be completed within about 4 weeks, though this is very flexible.
* Please let me know if any of you require some more time.
* We can also use @editorialbot to set automatic reminders if you know you'll be away for a known period of time.

There are links to the JOSS reviewer guidelines at the top of this issue. Please feel free to ping me here (@leouieda) or email me privately if you have any questions/concerns.
