### 2022-03-22 #4217 @hfmark
Hi @valentineap - here are some more comments for you! I'll try not to be too repetitive, but I've got some overlap with @hemmelig's thoughts (plus a few other small things). Overall, I think this is a really neat tool, and as someone who has spent far too much time trying to compile and run reflectivity codes, I appreciate how easy it is to install and run.

**Code functionality**

I'm also a bit concerned about the test failing. I wonder if the finite difference approximation for phi is missing a 1/r term somewhere? Seems more likely to be something coordinate-related like that than an actual failure of the code.

**Documentation**

It doesn't seem like you anticipate adding much to this package in the future, but I'd also endorse a line to the readme inviting people to (a) contact the authors or (b) submit a pull request if they want to contribute to the package.

The example provided in the documentation under "Calculating seismic observables" places the receivers on the water layer, which the code doesn't support. I know the next doc page has a working and fully annotated example, but it can't hurt to have this one also be fully functional.

Installation is smooth and the dependencies are, as promised, minimal. Still, it would be good to have a bulleted list of those in the readme and/or the documentation, including any "optional" dependencies like tqdm and matplotlib.

**The paper**

The "Statement of need" strongly emphasizes `pyprop8`'s potential as a simple, lightweight tool for teaching and testing, but the "Applications" section lists some real-world research applications of the underlying methodology. When you say that "The computational approach described by O'Toole and Woodhouse (2011) (but not the pyprop8 implementation)" has been used for these things, does that mean that `pyprop8` is implementing a simplified version of the O'Toole and Woodhouse formulation, or just that the original `prop8` was used instead of this particular code? I'm guessing the latter, but given how strongly the Statement of need emphasizes teaching/outreach uses rather than research, it would be good to clarify that the methodology is fully implemented and the fact that it's easy to use is a nice bonus.

**A suggestion** which you can take or leave:

Any thoughts on adding some sort of `LayeredStructureModel.earth_flattening_transform()` method? For body waves, at least, it's easy to do (e.g. Chapman and Orcutt 1985), and it could give users an option for working with slightly larger ranges.

Most of these comments are fairly minor, but I'll put some of the small documentation things in a pull request and open an issue for the test failure since that seems to be the preferred way for JOSS.
