### 2022-03-22 #4217 @valentineap
@hemmelig @hfmark Thank you both for taking the time to do this, and for your detailed comments -- much appreciated. I will take a few days to digest before I respond to anything.

@hemmelig The other reference for microseismic applications would be Tom O'Toole's PhD thesis: record [here](https://ora.ox.ac.uk/objects/uuid:9dcaca2e-c141-4e8a-94f6-34a9ff50d061). IIRC, he had an unpublished chapter looking at Preese Hall fracking. Unfortunately it does not seem to be publicly available in anything other than dead-tree format.

@leouieda How does this work on a technical level -- i.e. how do you want me to keep track of what I've done to address each comment?
