### 2022-03-29 #4217 @leouieda
@hemmelig @hfmark thank you for the detailed reviews and suggestions! 🎉

> I have some further notes, but I’m not 100% sure how exactly to proceed. I suspect it would be easiest if I opened a pull request and referenced it to this review – would that be correct, @leouieda? These primarily expand on some of the notes already made above.

Yes, you are definitely encouraged to open issues and PRs (if you're feeling generous) directly in the source repository.

> How does this work on a technical level -- i.e. how do you want me to keep track of what I've done to address each comment?

@valentineap my recommendation would be for you or the reviewers to open issues for each of the tasks outlined in the reviews (fix the failing test, expanding the example documentation, creating community guidelines, code formatting, documenting internal functions, edits to the paper, etc).

**When opening issues or PRs, please include `openjournals/joss-reviews#4217` in the description so that it links back to the review and is more easily tracked.**

> It doesn't seem like you anticipate adding much to this package in the future

Regardless, we still require the minimum community guidelines to be present for acceptance. See https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines. Personally, I would recommend adding 2 files:

* `CONTRIBUTING.md` explaining the channels of communication (where to open bug reports, ask questions, etc) and minimal information about contributing (which branch to make PRs against, how to run the tests, how to build the docs, etc). For example, this is what I use for my own projects: https://github.com/fatiando/community/blob/main/CONTRIBUTING.md (of course, we don't require that much information for JOSS)
* `CODE_OF_CONDUCT.md` with the CoC for the project. A good starting point is the [Contributor Covenant](https://www.contributor-covenant.org/).

@valentineap please keep me and the reviewers informed of the progress and feel free to ping us in issues/PRs if needed. As I said, the JOSS review is usually interactive with all of us working towards making sure the software meets the requirements.
