### 2022-04-28 valentineap/pyprop8#7 @valentineap
Provide exemplar teaching material
Comment from @hemmelig in openjournals/joss-reviews#4217

> I feel it might see far greater uptake if a more fleshed out example were provided, perhaps in the form of a Jupyter Notebook. These can be freely hosted online (for example, using MyBinder) and would bypass the need for any given user to install the package before they try it out. Equally, this may just be my relative naivety of the sorts of things one might use this for.
>
>  Consider providing a fleshed out Jupyter Notebook example demonstrating the value of the package for educational purposes, as stated in the paper.
