### 2022-04-28 valentineap/pyprop8#8 @valentineap
Create contributor guidelines
From @leouieda in openjournals/joss-reviews#4217

> Regardless, we still require the minimum community guidelines to be present for acceptance. See https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines. Personally, I would recommend adding 2 files:
>
> CONTRIBUTING.md explaining the channels of communication (where to open bug reports, ask questions, etc) and minimal information about contributing (which branch to make PRs against, how to run the tests, how to build the docs, etc). For example, this is what I use for my own projects: https://github.com/fatiando/community/blob/main/CONTRIBUTING.md (of course, we don't require that much information for JOSS)
> CODE_OF_CONDUCT.md with the CoC for the project. A good starting point is the [Contributor Covenant](https://www.contributor-covenant.org/).
