### 2022-04-28 valentineap/pyprop8#9 @valentineap
Check comparisons with prop8 paper
Comment from @hemmelig in openjournals/joss-reviews#4217:
> I ran the provided script for re-producing the figures presented in O’Toole & Woodhouse (2011) and O’Toole et al. (2012). Critically, these only run if the user installs matplotlib, which is not addressed anywhere in the documentation. Where possible, I tried to ensure I was comparing the figures at a similar scale/aspect ratio. By eye, the figures were largely identical. I note, however, some differences:
>
> 2011, Fig 1 – vertical traces show a slight concave deflection after the primary peak (strongest in proximal traces). There also appear to be a number of deflections later in the prop8 time series that are absent in the pyprop8 figures - e.g., using the notation “r2” to represent the nth trace (from top to bottom) of the radial component: r2, r4, r5, r10, t2, t6, v3, v7)
> 2011, Fig 5 – pyprop8 figure seems significantly smoother than that presented in the the original paper
> 2012, Fig 1 – a number of pyprop8 the GPS excitation kernels in pyprop8 do not appear to meet the reference line as they do in the original paper figure – e.g. Mr,theta East/Vertical, etc. This may well just be a scaling/reference point issue, but would be good to get some clarity.
> 2012, Fig 2 – a much clearer example of what I noted for Fig 1. Depth, zc East component intersects with the reference line, but the pyprop8 example does not. Time, tc North component looks quite different, too. Latitude North/Vertical, same note.

Also deal with matplotlib dependency.
