### 2022-04-28 valentineap/pyprop8#11 @valentineap
Fix documentation example
Comment from @hfmark in openjournals/joss-reviews#4217:

> The example provided in the documentation under "Calculating seismic observables" places the receivers on the water layer, which the code doesn't support. I know the next doc page has a working and fully annotated example, but it can't hurt to have this one also be fully functional.
