### 2022-04-28 valentineap/pyprop8#12 @valentineap
Clarify paper 
Comment from @hfmark in openjournals/joss-reviews#4217:

> When you say that "The computational approach described by O'Toole and Woodhouse (2011) (but not the pyprop8 implementation)" has been used for these things, does that mean that pyprop8 is implementing a simplified version of the O'Toole and Woodhouse formulation, or just that the original prop8 was used instead of this particular code? I'm guessing the latter, but given how strongly the Statement of need emphasizes teaching/outreach uses rather than research, it would be good to clarify that the methodology is fully implemented and the fact that it's easy to use is a nice bonus.
