### 2022-04-28 valentineap/pyprop8#13 @valentineap
Earth-flattening transform
Comment from @hfmark in openjournals/joss-reviews#4217:

> Any thoughts on adding some sort of LayeredStructureModel.earth_flattening_transform() method? For body waves, at least, it's easy to do (e.g. Chapman and Orcutt 1985), and it could give users an option for working with slightly larger ranges.
