### 2022-04-28 valentineap/pyprop8#14 @valentineap
Tidy up tests
As well as fixing the apparent error (valentineap/pyprop8#6) , tidy up/document the testing code. Comment from @hemmelig in openjournals/joss-reviews#4217:

> These should be a lot easier to run, and would be good to encourage the users to run them and provide an explanation. They also do not appear to work – specifically, the derivatives calculated for perturbations to phi are quite out. I note as well that the delta scaling is different for phi than for r or z.
>
> Also, is there a reason why the setup for the test event is identical to the O’Toole & Woodhouse (2011) paper example, but at 20 km rather than the 34 km used in the paper?
