### 2022-04-28 #4217 @valentineap
@hemmelig @hfmark @leouieda Many thanks again for all your comments above. Apologies for being slow to respond. Hopefully I will have a bit more time in the next couple of weeks to make some progress on this.

As you will see I have created a number of issues that I hope capture the salient points of your reviews. Please let me know if you think I have missed something (or just create additional issues!). I will add any comments in response and document the actions taken within the issue tracker.
