### 2022-04-28 valentineap/pyprop8#7 @valentineap
@hemmelig I'm not sure what the right thing to do here is. On the one hand, I see the appeal of making materials available and making it easy for people to 'try before they buy'. On the other: I feel that a set of teaching materials using `pyprop8` is a distinct entity from the code itself, and I'm not sure that mixing the two is necessarily a good long-term strategy. I'm also not sure how useful teaching materials will really be if they are conceived without the context of a particular course/audience/etc.

Perhaps a half-way house would be to provide some of the existing examples (e.g. reproducing the O'Toole & Woodhouse figures) in a format that enables users to immediately run them with MyBinder. I had also wondered about using something like Streamlit to make some interactive widgets (similar to what we did here: https://share.streamlit.io/charlesll/i-melt/imelt_streamlit.py).

I'd welcome any further comments/suggestions you may have on this. I'll tag @hfmark and @leouieda too in case they wish to weigh in.
