### 2022-04-28 a1fe81f
Fix to handle non-zero origin times
