+++ src/pyprop8/_core.py
@@ -1,27 +1,31 @@
-
 import numpy as np
 from pyprop8 import _scaledmatrix as scm
 import scipy.special as spec
-#from tqdm.autonotebook import tqdm edited by MS 2/3/21 due to warning in Anaconda JupyterLab
+
+# from tqdm.autonotebook import tqdm edited by MS 2/3/21 due to warning in Anaconda JupyterLab
 import warnings
+
 try:
     from tqdm import tqdm
 except ImportError:
     print("Unable to import 'tqdm'; progress bars will not be shown.")
     # Create a do-nothing implementation
     class tqdm:
-        def __init__(self,*args,**kwargs):
+        def __init__(self, *args, **kwargs):
             pass
-        def update(self,i):
+
+        def update(self, i):
             pass
+
         def close(self):
             pass
 
 
-PLANETARY_RADIUS=6371.
+PLANETARY_RADIUS = 6371.0
 
-def gc_dist(lat1,lon1,lat2,lon2,radius=PLANETARY_RADIUS):
-    '''
+
+def gc_dist(lat1, lon1, lat2, lon2, radius=PLANETARY_RADIUS):
+    """
     Calculate the great-circle distance between two points on the surface of a
     sphere.
 
@@ -33,15 +37,17 @@ def gc_dist(lat1,lon1,lat2,lon2,radius=PLANETARY_RADIUS):
 
     :return: The great-circle distance (km).
     :rtype: float
-    '''
-    dlat = lat2-lat1
-    dlon = lon2-lon1
+    """
+    dlat = lat2 - lat1
+    dlon = lon2 - lon1
 
-    u = np.sin(dlat/2)**2 + np.cos(lat1)*np.cos(lat2)*np.sin(dlon/2)**2
-    dist = 2*radius*np.arcsin(np.sqrt(u))
+    u = np.sin(dlat / 2) ** 2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon / 2) ** 2
+    dist = 2 * radius * np.arcsin(np.sqrt(u))
     return dist
-def gc_azimuth(lat1,lon1,lat2,lon2):
-    '''
+
+
+def gc_azimuth(lat1, lon1, lat2, lon2):
+    """
     Azimuth cw from North to go *to* (lat2,lon2) from (lat1,lon1)
 
     :param float lat1: Latitude of point 1
@@ -52,14 +58,17 @@ def gc_azimuth(lat1,lon1,lat2,lon2):
     :return: The azimuth (radians).
     :rtype: float
 
-    '''
-    dlon = lon2-lon1
-    return np.arctan2(np.cos(lat2)*np.sin(dlon),np.cos(lat1)*np.sin(lat2)-np.sin(lat1)*np.cos(lat2)*np.cos(dlon))
+    """
+    dlon = lon2 - lon1
+    return np.arctan2(
+        np.cos(lat2) * np.sin(dlon),
+        np.cos(lat1) * np.sin(lat2) - np.sin(lat1) * np.cos(lat2) * np.cos(dlon),
+    )
 
 
 class PointSource:
-    def __init__(self,x,y,dep,Mxyz,F,time):
-        '''
+    def __init__(self, x, y, dep, Mxyz, F, time):
+        """
         Create an object representing a seismic source. Source mechanism
         represented as the sum of a moment tensor and a force, acting at a
         single point in space. Alternatively N distinct moment tensor/force
@@ -78,33 +87,40 @@ class PointSource:
         :param float or datetime.datetime time: The event time, expressed either
             as an instance of :py:class:`datetime.datetime`, or as seconds
             relative to some arbitrary epoch.
-        '''
+        """
         self.x = x
         self.y = y
         self.dep = dep
[…]
