+++ src/pyprop8/_scaledmatrix.py
@@ -1,205 +1,287 @@
 import numpy as np
 
+
 class ScaledMatrixStack:
-    def __init__(self,data=None,scale=None,nStack=None,N=None,M=None,name=None,copy=False,dtypeData='float64',dtypeScale='float64'):
+    def __init__(
+        self,
+        data=None,
+        scale=None,
+        nStack=None,
+        N=None,
+        M=None,
+        name=None,
+        copy=False,
+        dtypeData="float64",
+        dtypeScale="float64",
+    ):
         self.name = name
         if data is None:
-            if scale is not None: raise ValueError("Optional argument 'scale' cannot be provided without setting 'data'.")
-            if nStack is None or N is None or M is None: raise ValueError("ScaledMatrixStack dimensions must be specified")
+            if scale is not None:
+                raise ValueError(
+                    "Optional argument 'scale' cannot be provided without setting 'data'."
+                )
+            if nStack is None or N is None or M is None:
+                raise ValueError("ScaledMatrixStack dimensions must be specified")
             self.nStack = nStack
-            self.shape = (N,M)
-            self.M = np.zeros([nStack,N,M],dtype=dtypeData)
-            self.scale = np.zeros([nStack],dtype=dtypeScale)
+            self.shape = (N, M)
+            self.M = np.zeros([nStack, N, M], dtype=dtypeData)
+            self.scale = np.zeros([nStack], dtype=dtypeScale)
         else:
-            if not (nStack is None and N is None and M is None): raise ValueError("Do not provide dimensions in conjunction with data")
+            if not (nStack is None and N is None and M is None):
+                raise ValueError("Do not provide dimensions in conjunction with data")
             try:
-                self.nStack,N,M = data.shape
-                self.shape = (N,M)
+                self.nStack, N, M = data.shape
+                self.shape = (N, M)
             except AttributeError:
-                raise TypeError("ScaledMatrixStack 'data' argument has invalid type (%s)"%type(data))
+                raise TypeError(
+                    "ScaledMatrixStack 'data' argument has invalid type (%s)"
+                    % type(data)
+                )
             except ValueError:
-                raise TypeError("ScaledMatrixStack 'data' argument appears to have wrong shape")
+                raise TypeError(
+                    "ScaledMatrixStack 'data' argument appears to have wrong shape"
+                )
             if copy:
                 self.M = data.copy()
                 if scale is None:
-                    self.scale = np.zeros(self.nStack,dtype=dtypeScale)
+                    self.scale = np.zeros(self.nStack, dtype=dtypeScale)
                 else:
                     try:
-                        if scale.shape!=(self.nStack,):
-                            raise ValueError("ScaledMatrixStack 'scale' argument does not have expected shape")
+                        if scale.shape != (self.nStack,):
+                            raise ValueError(
+                                "ScaledMatrixStack 'scale' argument does not have expected shape"
+                            )
                         self.scale = scale.copy()
-                    except AttributeError: # Assume it's a scalar
-                        self.scale = np.full((self.nStack,),scale)
+                    except AttributeError:  # Assume it's a scalar
+                        self.scale = np.full((self.nStack,), scale)
             else:
                 self.M = data
                 try:
-                    if scale.shape!=(self.nStack,):
[…]
