### 2022-04-28 b291c38
Moved backend propagation code to separate file
