+++ src/pyprop8/_core.py
@@ -1,5 +1,5 @@
 import numpy as np
-from pyprop8 import _scaledmatrix as scm
+from pyprop8._propagators import *
 import scipy.special as spec
 
 # from tqdm.autonotebook import tqdm edited by MS 2/3/21 due to warning in Anaconda JupyterLab
@@ -1407,234 +1407,6 @@ def compute_spectra(
             return spectra[:, :, ss, :, :], d_spectra[:, :, ss, :, :, :]
 
 
-def compute_H_matrices(k, omega, dz, sigma, mu, rho, isrc, irec, do_derivatives=False):
-    """
-    Compute the "H" matrices, as defined in O'Toole & Woodhouse 2011.
-    """
-    # Note on derivatives: the P-SV calculations are done with everything
-    # propagated to the receiver depth, whereas SH calculations mix source
-    # and receiver quantities. Thus the algorithmic structure of the derivative
-    # calculation differs.
-    nlayers = dz.shape[0]
-    # Propagate surface b/c to receiver
-    if do_derivatives:
-        surface_bc_sh_drv = []
-        surface_bc_psv_drv = []
-    if mu[0] == 0:
-        surface_bc_sh = oceanFloorBoundary(dz[0], omega, k, sigma[0], rho[0], True)
-        surface_bc_psv = oceanFloorBoundary(dz[0], omega, k, sigma[0], rho[0], False)
-        ibc = 1
-        if do_derivatives:
-            surface_bc_sh_drv += [
-                oceanFloorBoundary_deriv(dz[0], omega, k, sigma[0], rho[0], True)
-            ]
-            surface_bc_psv_drv += [
-                oceanFloorBoundary_deriv(dz[0], omega, k, sigma[0], rho[0], False)
-            ]
-    else:
-        surface_bc_sh = freeSurfaceBoundary(k.shape[0], True)
-        surface_bc_psv = freeSurfaceBoundary(k.shape[0], False)
-        ibc = 0
-
-    for i in range(ibc, irec):
-        # Deal with derivatives first - before we do in-place propagation of the vectors
-        if do_derivatives:
-            for j in range(len(surface_bc_sh_drv)):
-                # Normal propagation on every derivative that already exists
-                surface_bc_sh_drv[j], _, surface_bc_psv_drv[j] = propagate(
-                    omega,
-                    k,
-                    -dz[i],
-                    sigma[i],
-                    mu[i],
-                    rho[i],
-                    m2=surface_bc_sh_drv[j],
-                    m6=surface_bc_psv_drv[j],
-                )
-            sh_drv, _, psv_drv = propagate_deriv(
-                omega,
-                k,
-                -dz[i],
-                sigma[i],
-                mu[i],
-                rho[i],
-                m2=surface_bc_sh,
-                m6=surface_bc_psv,
-                inplace=False,
-            )
-            surface_bc_sh_drv += [sh_drv]
-            surface_bc_psv_drv += [psv_drv]
[…]
