### 2022-04-28 efb6a5c
Consolidated class definitions in code
