### 2022-04-28 valentineap/pyprop8#10 @valentineap
- [x] Auto-format with Black (#15)
- [ ] Set up Github workflow to enforce Black on commits?
- [x] Break `_core.py` into two parts: "things that the user will touch" (still in `_core.py`) and everything else (`_propagtors.py`) (#15)
- [ ] Documentation of internal routines
- [ ] Link to paper equations
- [ ] Plotting?
