### 2022-05-06 bd025aa
Partial solution to #6
