+++ src/pyprop8/tests.py
@@ -14,7 +14,7 @@ def test_spatial_derivatives(
     pad_frac=1,
     derivatives=None,
     delta=1e-4,
-    delta_scales={"depth": 1, "phi": 1e-2, "r": 1},
+    delta_scales={"depth": 1, "phi": 1e-4, "r": 1},
     source_time_function=None,
 ):
     """
@@ -74,6 +74,7 @@ def test_spatial_derivatives(
         pad_frac=pad_frac,
         derivatives=derivatives,
         source_time_function=source_time_function,
+        xyz=False,
     )
     assert len(seis0.shape) == nDimSta + nDimChannels + nDimTime
     assert len(drv.shape) == nDimSta + nDimChannels + nDimTime + nDimDerivs
@@ -93,6 +94,7 @@ def test_spatial_derivatives(
             pad_frac=pad_frac,
             derivatives=None,
             source_time_function=source_time_function,
+            xyz=False,
         )
         fd = (seis - seis0) / (delta * delta_scales["r"])
         err = drv[deriv_comp(derivatives.i_r)] - fd
@@ -117,8 +119,9 @@ def test_spatial_derivatives(
             pad_frac=pad_frac,
             derivatives=None,
             source_time_function=source_time_function,
+            xyz=False,
         )
-        fd = (seis - seis0) / (delta * delta_scales["phi"])
+        fd = (seis - seis0)*(180/np.pi) / (delta * delta_scales["phi"])
         err = drv[deriv_comp(derivatives.i_phi)] - fd
         if nDimTime == 0:
             norm = abs(drv[deriv_comp(derivatives.i_phi)])
@@ -140,6 +143,7 @@ def test_spatial_derivatives(
             pad_frac=pad_frac,
             derivatives=None,
             source_time_function=source_time_function,
+            xyz=False,
         )
         fd = (seis - seis0) / delta
         err = drv[deriv_comp(derivatives.i_dep)] - fd
