### 2022-05-13 b18e397
Added missing term to phi derivative in rotated (x/y/z) case. Largely addresses Issue #6.
