### 2022-05-30 7188135
Started post-install tests script
