### 2022-06-01 2bc8fde
Renamed 'depth' derivative to become 'z' derivative to avoid potential for confusion. This derivative represents the change in seismogram due to a *positive* change in the z-coordinate of the seismic source, which is equivalent to a *decrease* in the source depth.
