### 2022-06-20 e56be47
Fixed some issues with rotation of static displacement field derivatives
