+++ src/pyprop8/tests.py
@@ -1,165 +1,11 @@
-import numpy as np
 import pyprop8 as pp
-from pyprop8.utils import stf_trapezoidal, rtf2xyz, make_moment_tensor
-from collections import defaultdict
+from pyprop8.utils import stf_trapezoidal,make_moment_tensor,rtf2xyz
+import numpy as np
 
-
-def test_spatial_derivatives(
-    model,
-    source,
-    stations,
-    nt=257,
-    dt=0.5,
-    alpha=0.023,
-    pad_frac=1,
-    derivatives=None,
-    delta=1e-4,
-    delta_scales={"depth": 1, "phi": 1e-4, "r": 1},
-    source_time_function=None,
-    xyz=True,
-):
-    """
-    Compute finite-difference derivatives and compare with the 'analytic'
-    derivatives generated by pyprop8. All arguments are as required by
-    `pyprop8.compute_seismograms()` except:
-
-    delta        - float, perturbation for finite-difference calculation
-    delta_scales - dict, containing keys that correspond to the keywords
-                   available when creating a `pyprop8.DerivativeSwitches()`
-                   object, defining a parameter-specific rescaling for `delta`
-                   to account for the fact that parameters are specified in
-                   dimensionalised units. If key is not found, 'delta' is used
-                   unmodified.
-
-
-    """
-    print("Testing spatial derivatives against finite difference calculation...")
-    if source.nsources > 1:
-        raise ValueError("test_derivatives requires a single force/mechanism")
-    if source_time_function is None:
-        source_time_function = lambda w: stf_trapezoidal(w, 3, 6)
-    if derivatives is None:
-        derivatives = pp.DerivativeSwitches(r=True, phi=True, z=True)
-    if type(delta_scales) is not defaultdict:
-        # Make delta_scales have a default value of '1'
-        delta_scales = defaultdict(lambda: 1, **delta_scales)
-
-    nDimSta = stations.nDim
-    if derivatives.nderivs == 0:
-        raise ValueError("All derivatives are 'switched off'!")
-    elif derivatives.nderivs > 1:
-        nDimDerivs = 1
-    else:
-        nDimDerivs = 0
-    nDimChannels = 1
-    if nt > 1:
-        nDimTime = 1
-    else:
-        nDimTime = 0
-
-    # Function to handle slicing of derivative arrays regardless of shapes
-    deriv_comp = lambda sl: tuple(
-        nDimSta * [slice(None)]
-        + nDimDerivs * [sl]
-        + nDimChannels * [slice(None)]
-        + nDimTime * [slice(None)]
-    )
-
-    tt, seis0, drv = pp.compute_seismograms(
-        model,
-        source,
-        stations,
-        nt,
-        dt,
-        alpha,
-        pad_frac=pad_frac,
-        derivatives=derivatives,
-        source_time_function=source_time_function,
-        xyz=xyz,
-    )
-    assert len(seis0.shape) == nDimSta + nDimChannels + nDimTime
-    assert len(drv.shape) == nDimSta + nDimChannels + nDimTime + nDimDerivs
-    drverrs = []
-
-    if derivatives.r:
-        sta_pert = stations.copy()
-        sta_pert.rmin += delta * delta_scales["r"]
-        sta_pert.rmax += delta * delta_scales["r"]
-        tt, seis = pp.compute_seismograms(
-            model,
-            source,
-            sta_pert,
-            nt,
-            dt,
-            alpha,
-            pad_frac=pad_frac,
-            derivatives=None,
[…]
