+++ src/pyprop8/tests.py
@@ -1,9 +1,10 @@
 import pyprop8 as pp
-from pyprop8.utils import stf_trapezoidal,make_moment_tensor,rtf2xyz
+from pyprop8.utils import stf_trapezoidal, make_moment_tensor, rtf2xyz
 import numpy as np
 
+
 def tests():
-    print("Running tests. Using `pyprop8` from: %s"%pp.__file__)
+    print("Running tests. Using `pyprop8` from: %s" % pp.__file__)
     print("")
     print(" 1. Creating objects")
     model = pp.LayeredStructureModel(
@@ -24,7 +25,9 @@ def tests():
         np.zeros([3, 1]),
         0,
     )
-    stations = pp.RegularlyDistributedReceivers(30, 100, 7, 0, 360, 10, depth=3).asListOfReceivers()
+    stations = pp.RegularlyDistributedReceivers(
+        30, 100, 7, 0, 360, 10, depth=3
+    ).asListOfReceivers()
 
     derivs = pp.DerivativeSwitches(x=True, y=True, z=True)
 
@@ -32,10 +35,10 @@ def tests():
 
     print(" 2. Computing seismograms...")
 
-    nt=33 #257
-    dt=0.5
-    alpha=0.023
-    pad_frac=1
+    nt = 33  # 257
+    dt = 0.5
+    alpha = 0.023
+    pad_frac = 1
 
     tt, seis0, drv = pp.compute_seismograms(
         model,
@@ -52,13 +55,14 @@ def tests():
 
     epsilon = 1e-4
     print(" 3. Comparing with finite-difference derivatives.")
-    print("    Using finite-difference perturbation eps = %.2f metres"%(epsilon*1000))
+    print(
+        "    Using finite-difference perturbation eps = %.2f metres" % (epsilon * 1000)
+    )
     print("    a. Perturbing source in x.")
-
 
     source_x = source.copy()
     source_x.x += epsilon
-    
+
     tt, seis_x = pp.compute_seismograms(
         model,
         source_x,
@@ -72,15 +76,22 @@ def tests():
         xyz=True,
     )
 
-    fd = (seis_x - seis0)/epsilon # finite difference estimate
-    max_x = abs(drv[:,derivs.i_x,:,:]).max(-1).reshape(stations.nstations,3,1) # Maximum absolute value of trace
-    perc_err_x = 100*(abs(drv[:,derivs.i_x,:,:]-fd)/max_x) #For each trace, calculate finite difference 'error' as percentage of trace amplitude
-    print("       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"%perc_err_x.max())
+    fd = (seis_x - seis0) / epsilon  # finite difference estimate
+    max_x = (
+        abs(drv[:, derivs.i_x, :, :]).max(-1).reshape(stations.nstations, 3, 1)
+    )  # Maximum absolute value of trace
+    perc_err_x = 100 * (
+        abs(drv[:, derivs.i_x, :, :] - fd) / max_x
+    )  # For each trace, calculate finite difference 'error' as percentage of trace amplitude
+    print(
+        "       Worst-case difference between 'true' and finite-difference derivatives: %.3f%%"
+        % perc_err_x.max()
+    )
 
     print("    b. Perturbing source in y.")
     source_y = source.copy()
     source_y.y += epsilon
-    
+
     tt, seis_y = pp.compute_seismograms(
         model,
         source_y,
@@ -94,16 +105,22 @@ def tests():
         xyz=True,
     )
 
-    fd = (seis_y - seis0)/epsilon # finite difference estimate
-    max_y = abs(drv[:,derivs.i_y,:,:]).max(-1).reshape(stations.nstations,3,1) # Maximum absolute value of trace
[…]
