### 2022-06-22 valentineap/pyprop8#9 @valentineap
Unfortunately, but perhaps inevitably, there are various details that are not fully and unambiguously described in the papers. For example, the source-time function used in Fig. 1 of O'Toole & Woodhouse (2011) is described as a 'symmetrical cosine tapered boxcar', but no explicit formula is given, and I do not seem to have any scripts/code that can confirm precisely what was done. It would seem some information is missing, regarding the length of taper applied to the boxcar. [In fact the pyprop8 version was using a trapezoidal source time-function here, but changing this does not seem to resolve the issue.]

For that Fig.1, specifically, it seems to me that there are artefacts present in the traces shown in the paper: there are acausal long-period features that I suspect are 'edge effects' from some aspect of the processing. This is most evident in trace r17 (per @hemmelig's notation) where there is a clear downwards deflection in the first 30s of the trace, before the arrival of the main wavefront. Other traces including r2, r5 and r10 show something similar towards the end of the time series, as noted by the reviewer. Given that these do not display any 'move-out' or other coherent behaviour across traces, and have no obvious explanation within the simple physics of the setup here, I am comfortable not to be reproducing these in our new version.
