### 2022-06-23 43a9cbd
Reproduce 2012 paper
