### 2022-06-23 0774bd5
Updated docstrings
