+++ src/pyprop8/_scaledmatrix.py
@@ -1,7 +1,28 @@
 import numpy as np
 
+"""
+This module implements operations on exponentially-scaled matrices. This allows
+stable computation of matrix-matrix and matrix-vector products in cases where elements may
+be very large and/or very small.
+
+A general matrix M is represented as (s, A) so that M = exp(s). A 
+
+Further, this module implements computations for a 'stack' of matrices: this is simply a
+structure to enable efficient evaluation of many similar calculations. If, for example
+
+s1 = Stack{ A, B, C}   and s2 = Stack{ u, v, w}
+
+then
+
+s1 @ s2 = Stack{ A @ u , B @ v, C @ w}
+"""
 
 class ScaledMatrixStack:
+    """
+    A class to represent a collection (`stack`) of N x M matrices.
+    An individual matrix A is stored in the form (s, D) such that A = exp(s).D
+    where s is a scalar and D is an N x M matrix.    
+    """
     def __init__(
         self,
         data=None,
@@ -9,12 +30,34 @@ class ScaledMatrixStack:
         nStack=None,
         N=None,
         M=None,
-        name=None,
         copy=False,
         dtypeData="float64",
         dtypeScale="float64",
     ):
-        self.name = name
+        """
+        Initialise a stack (collection) of exponentially scaled matrices: each matrix A 
+        is expressed as (s, D) where A=exp(s).D. Usage is either:
+
+        scm = ScaledMatrixStack(data, scale, copy=False)
+
+        to populate the stack using pre-existing data and scale, or 
+
+        scm = ScaledMatrixStack(nStack, N, M, dtypeData="float64", dtypeScale="float64")
+        
+        to create an empty (zeros) stack of a given dimension and data type. 
+
+        Inputs:
+        data - array, shape (nStack, N, M): a set of nStack (N x M) matrices, {D1, D2, ...}
+        scale - array, shape (nStack,): the exponential scale factor for each of the
+                nStack matrices, {s1, s2, ...}. (Use scale = np.zeros(...) if `data`
+                is currently unscaled.
+        nStack - integer. The number of matrices in the stack.
+        N, M  - integers. The dimensions of individual matrices.
+        copy - True/False. If True, call .copy() on both `data` and `scale`; if False,
+                use the versions as passed in.
+        dtypeData - any valid `numpy.dtype` specification. The data type used for `D`.
+        dtypeScale - any valid `numpy.dtype` specification. The data type used for `s`.
+        """
         if data is None:
             if scale is not None:
                 raise ValueError(
@@ -69,6 +112,14 @@ class ScaledMatrixStack:
             # self.rescale()
 
     def copy(self, dest=None):
+        """Create a copy of a ScaledMatrixStack. 
+        ```
+        scm = ScaledMatrixStack(...)
+        new = scm.copy()
+        ```
+        """
+        "
+        
         if dest is None:
             return ScaledMatrixStack(self.M, self.scale, copy=True)
         else:
@@ -77,12 +128,26 @@ class ScaledMatrixStack:
             return dest
 
     def rescale(self):
+        """
+        Update the entries in a ScaledMatrixStack so that each matrix D
+        has unit maximum absolute value.
+        """
         mx = np.abs(self.M).max((1, 2))
         mx = np.where(mx > 0, mx, 1)
         self.M /= mx.reshape(-1, 1, 1)
         self.scale += np.log(mx)
 
     def __getitem__(self, key):
+        """
+        Enable indexing.
+        ```
+        scm = ScaledMatrixStack(...)
+        a = scm[3:7] # <--- Select examples 3-7 from the stack; return as a new stack
[…]
