### 2022-06-23 d822994
Added some references to explicit equations in papers
