+++ src/pyprop8/_propagators.py
@@ -5,6 +5,7 @@ from pyprop8 import _scaledmatrix as scm
 ### Boundary condition vectors ###
 ##################################
 def freeSurfaceBoundary(nk, sh=False):
+    # (2011) Eq. 85 (with P-SV system in minor vector form)
     if sh:
         m = np.zeros([nk, 2, 1], dtype="complex128")
     else:
@@ -14,6 +15,7 @@ def freeSurfaceBoundary(nk, sh=False):
 
 
 def oceanFloorBoundary(depth, omega, k, sigma, rho, sh=False):
+    # (2011) Eqs. 86 & 88 (P-SV in minor vector form)
     nk = k.shape[0]
     if sh:
         m = np.zeros([nk, 2, 1], dtype="complex128")
@@ -29,6 +31,7 @@ def oceanFloorBoundary(depth, omega, k, sigma, rho, sh=False):
 
 
 def oceanFloorBoundary_deriv(depth, omega, k, sigma, rho, sh=False):
+    # Derivative of (2011) Eqs. 86 & 88 wrt ocean depth (i.e. layer thickness)
     nk = k.shape[0]
     if sh:
         m = np.zeros([nk, 2, 1], dtype="complex128")
@@ -43,6 +46,7 @@ def oceanFloorBoundary_deriv(depth, omega, k, sigma, rho, sh=False):
 
 
 def underlyingHalfspaceBoundary(omega, k, sigma, mu, rho, sh=False):
+    # (2011) Eqs. 89 & 90
     nk = k.shape[0]
 
     zmu = np.lib.scimath.sqrt(k**2 - rho * omega**2 / mu)
@@ -52,6 +56,7 @@ def underlyingHalfspaceBoundary(omega, k, sigma, mu, rho, sh=False):
         m[:, 1, 0] = mu * zmu
     else:
         zsig = np.lib.scimath.sqrt(k**2 - rho * omega**2 / sigma)
+        # xi <-- (2011) Eq. 91
         xi = np.zeros_like(zsig)
         xi[k == 0] = np.sqrt(mu / sigma)
         xi[k > 0] = (rho * omega**2 / sigma - k[k > 0] ** 2 * (1 + mu / sigma)) / (
@@ -68,9 +73,10 @@ def underlyingHalfspaceBoundary(omega, k, sigma, mu, rho, sh=False):
 
 
 def sourceVector(MT, F, k, sigma, mu):
+    # (2011) Eqs. 21 & 22.
     nk = k.shape[0]
-    s = np.zeros([nk, 4, 5], dtype="complex128")
-    s2 = np.zeros([nk, 2, 5], dtype="complex128")
+    s = np.zeros([nk, 4, 5], dtype="complex128") # P-SV system, eq.21
+    s2 = np.zeros([nk, 2, 5], dtype="complex128") # SH system, eq. 22
     s[:, 0, 2] = MT[2, 2] / sigma
     s[:, 2, 2] = -F[2]
     s[:, 3, 2] = (
@@ -95,11 +101,13 @@ def sourceVector(MT, F, k, sigma, mu):
 
 
 def sourceVector_ddep(MT, F, omega, k, sigma, mu, rho):
+    # (2012) Eqs. A37 & A38
+    # i.e. derivative of (2011) Eqs. 21 & 22 wrt source depth.
     lam = sigma - 2 * mu
     gamma = mu * (3 * lam + 2 * mu) / (lam + 2 * mu)
     nk = k.shape[0]
-    s = np.zeros([nk, 4, 5], dtype="complex128")
-    s2 = np.zeros([nk, 2, 5], dtype="complex128")
[…]
