### 2022-06-24 26e38bb
Created contributor guidelines
