+++ docs/gettingstarted.rst
@@ -28,3 +28,13 @@ Some additional convenience functions are found in the module ``pyprop8.utils``.
 The package includes further modules that are used internally, but which are unlikely to be useful to end-users.
 
 For more details, read :ref:`walkthrough` and take a look at the :ref`example`.
+
+-------------------------
+Testing your installation
+-------------------------
+To verify that your installation is working correctly, you may wish to run the tests that are distributed as part of the package. To access these, enter the following in a Python interpreter::
+
+  from pyprop8 import tests
+  tests.tests()
+
+This will run and report a number of calculations.
