### 2022-06-25 f50c5db
Added explanation of equations
