### 2022-06-27 #4217 @leouieda
👋🏾 Hi @valentineap, just checking in to see how this is progressing. I see you've closed a lot of the open issues related to the review.

Just wanted to give a heads up that I'll be on off on vacation between 4-27 July and won't be available. If you feel like you can wrap this up in the meantime, I can try to arrange for another editor to finalise the review while I'm away. Otherwise, we can pick this up when I return.
