### 2022-07-21 valentineap/pyprop8#13 @valentineap
I've added a utility function to transform a model table, see https://github.com/valentineap/pyprop8/commit/9ce615e76f0169b8de5136d947298665ddb3bee7.
