### 2022-08-01 valentineap/pyprop8#7 @valentineap
I have now added a [Binder demo](https://mybinder.org/v2/gh/valentineap/pyprop8/HEAD?labpath=examples%2Fdemo.ipynb) [ecc50b46fb069681c50398ea8d9e169e537942e1]. With this and the other things in the examples folder I think there is enough for people to see what the code does and how it works. I have some things I hope to add over time, but this is always going to be open-ended and I'm keen to get the JOSS paper 'out there'.

I am therefore going to declare this issue 'closed', but feel free to reopen if you think there are particular things that still require demonstration.
