### 2022-08-01 valentineap/pyprop8#9 @valentineap
I think these comparisons are as close as they're going to get, and the differences are explainable. I have clarified this within the relevant scripts. I am therefore closing this issue, but please reopen if you still have concerns.
