+++ USERS.md
@@ -1,0 +1,11 @@
+# Users
+
+This file contains a list of [repositories](#repositories) and [publications](#publications) that use `pyprop8`. Please feel free to add your own work to the list.
+
+## Repositories
+
+- [waveform-ot](https://github.com/msambridge/waveform-ot/)
+
+## Papers
+
+- Sambridge, Jackson & Valentine (2022). Geophysical Inversion and Optimal Transport, *Geophysical Journal International*, [doi:10.1093/gji/ggac151](https://doi.org/10.1093/gji/ggac151)
