+++ USERS.md
@@ -6,6 +6,6 @@ This file contains a list of [repositories](#repositories) and [publications](#p
 
 - [waveform-ot](https://github.com/msambridge/waveform-ot/)
 
-## Papers
+## Publications
 
 - Sambridge, Jackson & Valentine (2022). Geophysical Inversion and Optimal Transport, *Geophysical Journal International*, [doi:10.1093/gji/ggac151](https://doi.org/10.1093/gji/ggac151)
