+++ USERS.md
@@ -8,4 +8,14 @@ This file contains a list of [repositories](#repositories) and [publications](#p
 
 ## Publications
 
-- Sambridge, Jackson & Valentine (2022). Geophysical Inversion and Optimal Transport, *Geophysical Journal International*, [doi:10.1093/gji/ggac151](https://doi.org/10.1093/gji/ggac151)
+- Sambridge, Jackson & Valentine (2022). Geophysical Inversion and Optimal Transport, *Geophysical Journal International*, doi:[10.1093/gji/ggac151](https://doi.org/10.1093/gji/ggac151)
+
+## Earlier studies
+
+In addition, the following papers rely on the seismogram synthesis approach described in O'Toole & Woodhouse (2011) and O'Toole et al. (2012), but **not** the `pyprop8` implementation:
+
+- Käufl, Valentine, O'Toole & Trampert (2014). A framework for fast probabilistic centroid&ndash;moment-tensor determination&mdash;inversion of regional static displacement measurements, *Geophysical Journal International*, doi:[10.1093/gji/ggt473](https://doi.org/10.1093/gji/ggt473).
+- Käufl, Valentine, de Wit & Trampert (2015). Robust and fast probabilistic source parameter estimation from near‐field displacement waveforms using pattern recognition, *Bulletin of the Seismological Society of America*, doi:[10.1785/0120150010](https://doi.org/10.1785/0120150010).
+- Käufl, Valentine, de Wit & Trampert (2016). Solving probabilistic inverse problems rapidly with prior samples. *Geophysical Journal International*, doi:[10.1093/gji/ggw108](https://doi.org/10.1093/gji/ggw108).
+- O'Toole, Valentine & Woodhouse (2013). Earthquake source parameters from GPS-measured static displacements with potential for real-time application, *Geophysical Research Letters*, doi:[10.1029/2012GL054209](https://doi.org/10.1029/2012GL054209).
+- O'Toole, Woodhouse, Verdon & Kendall (2013). Microseismic source mechanisms: What the waveform can tell us. *Proceedings of the  4th EAGE Passive Seismic Workshop*, doi:[10.3997/2214-4609.20142350](https://doi.org/10.3997/2214-4609.20142350).
