### 2022-08-01 d3ca75c
Update USERS.md
