### 2022-08-01 c3b7b9e
Update USERS.md
