### 2022-08-02 9518816
Trying to fix normalisation
