+++ examples/wavefield_movie.py
@@ -10,16 +10,16 @@ from mpl_toolkits.mplot3d import art3d as art3d
 import tqdm 
 
 
-lo_res = True
+lo_res = False
 if lo_res:
-    nt=30
-    delta_t = 2
-    ngrid=10
+    nt=60
+    delta_t = 1
+    ngrid=20
     ncontour=28
     dpi=100
-    framestep=10
-    nr = 5
-    nphi=16
+    framestep=5
+    nr = 10
+    nphi=36
 else:
     nt = 120
     delta_t = 0.5
@@ -29,7 +29,7 @@ else:
     framestep=1
     nr=50
     nphi=72
-
+box_size = 50
 
 model = pp.LayeredStructureModel([[ 3.00, 1.80, 0.00, 1.02],
                                   [ 2.00, 4.50, 2.40, 2.57],
@@ -41,7 +41,7 @@ strike = 180
 dip = 0
 rake = 0
 M0=1E7
-depth = 15
+depth = 5
 moment_tensor = rtf2xyz(make_moment_tensor(strike, dip, rake, M0, 0, 0))
 source = pp.PointSource(0,0,depth,moment_tensor,np.zeros([3,1]),0)
 stf = lambda w: stf_trapezoidal(w, 3, 6)
@@ -51,30 +51,37 @@ xx,yy = receivers.as_xy()
 tri = Triangulation(xx.flatten(),yy.flatten())
 
 
-time, seismograms = pp.compute_seismograms(model,source,receivers,nt,delta_t,source_time_function=stf)
+time, seismograms = pp.compute_seismograms(model,source,receivers,nt,delta_t,source_time_function=stf,stencil_kwargs={'kmin':0,'kmax':5,'nk':12000})
 
 fig = plt.figure(figsize=(8,8))
-ax1 = fig.add_axes((0.05,0.05,0.9,0.9),projection='3d')
+ax1 = fig.add_axes((0.0,0.0,1,1),projection='3d')
 ax2 = fig.add_axes((0.85,0.05,0.1,0.1),projection='3d')
 ax3 = fig.add_axes((0.05,0.05,0.1,0.1))
 
 
-grid_x,grid_y=np.meshgrid(np.linspace(-50,50,ngrid),np.linspace(-50,50,ngrid))
-cmax=abs(seismograms).max()*1.05
-scale_horiz=0.1
-scale_vert = 0.1
+grid_x,grid_y=np.meshgrid(np.linspace(-box_size,box_size,ngrid),np.linspace(-box_size,box_size,ngrid))
+
+
+max_norm = np.linalg.norm(seismograms,2,axis=2).max()
+
+scale_horiz = box_size/max_norm
+scale_vert = scale_horiz
+
+#cmax=abs(seismograms).max()*1.05
+cmax = max_norm*1.05
+#scale_horiz=0.01
+#scale_vert = 0.1
 view_azim_init = -45
 view_elev = 30
 nstatic = 120
 
 def animate(i):
+    # Plot the ground surface (with displacement)
     azim = view_azim_init
     if i>nstatic:
         azim = view_azim_init+i-nstatic
     ax1.azim = azim
     ax1.elev = view_elev
-    #view_azim=10
-    #view_elev=30
     ls = LightSource(90, 30)
     if i<nt:
         fx = CubicTriInterpolator(tri,seismograms[:,:,0,i].flatten())
@@ -97,21 +104,22 @@ def animate(i):
                            rcount=ngrid,ccount=ngrid)
     ax1.set_xlim(-50,50)
     ax1.set_ylim(-50,50)
-    ax1.set_xlabel("West-East")
[…]
