### 2022-08-03 f89430a
Improved docstrings
