+++ src/pyprop8/utils.py
@@ -2,12 +2,21 @@ import numpy as np
 
 
 def stf_trapezoidal(omega, trise, trupt):
-    """Trapezoidal source time function. This routine is a copy of one
-    in TBO'T's Matlab code, which is itself apparently derived from AXITRA.
-    The spectrum coded here appears to be the Fourier Transform of the
+    """
+    Trapezoidal source time function, in frequency domain. 
+    This routine is a copy of one in TBO'T's Matlab code, which is itself 
+    apparently derived from AXITRA.
+    
+    The spectrum coded here is the Fourier Transform of the
     convolution of two unit-area boxcars, one of which is non-zero in the
     range [0,trise], and the other of which is non-zero in the range
     [-trupt/2, trupt/2].
+
+    :param complex omega: Frequency at which evaluation is required (rad/s)
+    :param float trise: Rise time (s)
+    :param float trupt: Total duration of rupture (s)
+
+    :returns: Amplitude of source time function
     """
     uu = np.ones(omega.shape, dtype="complex128")
     uxx = np.ones_like(uu)
@@ -22,6 +31,19 @@ def stf_trapezoidal(omega, trise, trupt):
 
 
 def stf_cosine(omega, thalf):
+    """
+    Cosine source time function, in frequency domain.
+    
+    This function implements the Fourier transform of
+    f(t, T) = 1/(2T) (1+ cos( pi t / T))     -T < t < T
+            = 0                              otherwise
+    with T <--> ``thalf``
+
+    :param complex omega: Frequency at which evaluation is required (rad/s)
+    :param float thalf: Half-width of source (s)
+
+    :returns: Amplitude of source time function
+    """
     # This is the fourier transform of
     # f(t, T) = 1/(2T) (1+ cos( pi t / T))     -T < t < T
     #         = 0                              otherwise
@@ -33,6 +55,19 @@ def stf_cosine(omega, thalf):
     )
 
 def stf_boxcar(omega, thalf):
+    """
+    Boxcar source time function, in frequency domain.
+
+    This function implements the Fourier transform of 
+    f(t, T) = 1/(2T)   -T < t < T
+            = 0         otherwise
+    with T <--> ``thalf``
+
+    :param complex omega: Frequency at which evaluation is required (rad/s)
+    :param float thalf: Half-width of boxcar (s)
+
+    :returns: Amplitude of source time function
+    """ 
     # This is the fourier transform of
     # f(t, T) = 1/(2T)   -T < t < T
     #         = 0         otherwise
@@ -43,7 +78,16 @@ def stf_cosine_boxcar(omega, thalf, ratio=0.1):
     return np.pi**2 * np.cos(thalf*ratio*omega)*np.sin(thalf*omega*(ratio-1))/((ratio-1)*thalf*omega*(np.pi-2*ratio*thalf*omega)*(np.pi+2*ratio*thalf*omega))
 
 def clp_filter(w, w0, w1):
-    """Cosine low-pass filter"""
+    """
+    Cosine low-pass filter.
+    
+    :param complex w: Frequency a t which evaluation is required (rad/s)
+    :param float w0: Corner frequency for start of taper; at frequencies below this the filter has no impact.
+    :param float w1: Corner frequency for end of taper; frequencies above this are stopped.
+
+    :returns: Amplitude of the filter
+
+    """
     if w1 <= w0:
         raise ValueError(
             "clp_filter upper corner frequency must be greater than lower corner!"
@@ -56,8 +100,35 @@ def clp_filter(w, w0, w1):
         return 0
 
 
-def make_moment_tensor(strike, dip, rake, M0, eta, xtr):
-    """Construct moment tensor from strike/dip/rake"""
[…]
