+++ src/pyprop8/_propagators.py
@@ -5,6 +5,15 @@ from pyprop8 import _scaledmatrix as scm
 ### Boundary condition vectors ###
 ##################################
 def freeSurfaceBoundary(nk, sh=False):
+    """
+    Implement the boundary condition at a free surface.
+
+    :param int nk: Number of k-space samples
+    :param bool sh: True -> SH system; False -> P-SV system
+
+    :returns: An instance of :py:class:~pyprop8._scaledmatrix.ScaledMatrixStack: 
+        containing the boundary vector at each k value.
+    """
     # (2011) Eq. 85 (with P-SV system in minor vector form)
     if sh:
         m = np.zeros([nk, 2, 1], dtype="complex128")
@@ -15,6 +24,19 @@ def freeSurfaceBoundary(nk, sh=False):
 
 
 def oceanFloorBoundary(depth, omega, k, sigma, rho, sh=False):
+    """
+    Implement the boundary condition at the ocean floor
+
+    :param float depth: Water depth/ocean layer thickness
+    :param complex omega: Evaluation frequency
+    :param numpy.ndarray k: k-space wavenumber
+    :param float sigma: P-wave modulus
+    :param float rho: Density
+    :param bool sh: True -> SH system, False -> P-SV system
+
+    :returns: An instance of :py:class:~pyprop8._scaledmatrix.ScaledMatrixStack: 
+        containing the boundary vector at each k value.   
+    """
     # (2011) Eqs. 86 & 88 (P-SV in minor vector form)
     nk = k.shape[0]
     if sh:
@@ -31,6 +53,19 @@ def oceanFloorBoundary(depth, omega, k, sigma, rho, sh=False):
 
 
 def oceanFloorBoundary_deriv(depth, omega, k, sigma, rho, sh=False):
+    """
+    Implement the derivative of the ocean floor boundary vector wrt depth
+
+    :param float depth: Water depth/ocean layer thickness
+    :param complex omega: Evaluation frequency
+    :param numpy.ndarray k: k-space wavenumber
+    :param float sigma: P-wave modulus
+    :param float rho: Density
+    :param bool sh: True -> SH system, False -> P-SV system
+
+    :returns: An instance of :py:class:~pyprop8._scaledmatrix.ScaledMatrixStack: 
+        containing the derivative at each k value.
+    """
     # Derivative of (2011) Eqs. 86 & 88 wrt ocean depth (i.e. layer thickness)
     nk = k.shape[0]
     if sh:
@@ -46,6 +81,19 @@ def oceanFloorBoundary_deriv(depth, omega, k, sigma, rho, sh=False):
 
 
 def underlyingHalfspaceBoundary(omega, k, sigma, mu, rho, sh=False):
+    """
+    Implement the boundary condition for the underlying halfspace
+
+    :param complex omega: Evaluation frequency
+    :param numpy.ndarray k: k-space wavenumber 
+    :param float sigma: P-wave modulus
+    :param float mu: S-wave modulus
+    :param float rho: Density
+    :param bool sh: True -> SH system, False -> P-SV system
+
+    :returns: An instance of :py:class:~pyprop8._scaledmatrix.ScaledMatrixStack: 
+        containing the boundary vector at each k value.   
+    """
     # (2011) Eqs. 89 & 90
     nk = k.shape[0]
 
@@ -72,7 +120,22 @@ def underlyingHalfspaceBoundary(omega, k, sigma, mu, rho, sh=False):
     return scm.ScaledMatrixStack(m)
 
 
-def sourceVector(MT, F, k, sigma, mu):
+def sourceVector(MT, F, k, sigma, mu):    
+    """
+    Implement the source discontinity vector
+
[…]
