### 2022-08-08 5b0a50f
Removed visualisations.py from core package. The contents were rather limited in scope and not easily generalisable.
Instead, the examples/ directory now contains various visualisation
examples.

This has the added advantage that `matplotlib` is not required
to import any of the libraries in the core package.
