### 2022-08-08 6c875cf
Updated README to reflect changes in dependencies
