### 2022-08-08 e617088
Reorganising examples to allow per-example  files
