### 2022-08-08 92f5587
Improved figure
