+++ examples/idaho_wavefield_movie/idaho_wavefield_movie.py
@@ -1,0 +1,236 @@
+import numpy as np 
+import pyprop8 as pp
+from pyprop8.utils import rtf2xyz, make_moment_tensor,stf_trapezoidal
+import tqdm 
+from mayavi import mlab
+from scipy.interpolate import griddata
+
+print(pp.__file__)
+# This script generates an animated gif showing ground motion and
+# line-of-sight static displacement (~ InSAR) for an event
+
+nt = 401      # Number of time-steps in seismogram calculation
+delta_t = 0.1 # Time interval in seismogram calculation
+ngrid=560     # Number of grid points in Cartesian grid for interpolation
+nr=200         # Number of radii for seismogram calculation
+nphi=360       # Number of azimuths for seismogram calculation
+
+frame_file_fmt = 'animation/frame_%04i.png'
+
+# To ensure a high-quality resolution of the surface, increase the
+# spatial (k-space) sampling. Probably doesn't need to be this high!
+stencil_args = {'kmin':0,'kmax':7,'nk':20000}
+box_size = 65
+
+
+# Earth model to use
+# These values are Based on Crust1.0 for Idaho
+model = pp.LayeredStructureModel([( 16.5, 6.1, 3.6 , 2.7),
+                                 ( 14.6, 6.3, 3.7, 2.8),
+                                 (  6.4, 7.0, 4.0, 3.0),
+                                 (np.inf,7.9, 4.4, 3.3)])
+
+# Earthquake parameters
+# GCMT catalog values
+depth = 13.8
+moment_tensor_rtf = 1e6*np.array([[-2.320, 1.120, 1.150],
+                                  [ 1.120, 1.780,-5.970],
+                                  [ 1.150,-5.970, 0.535]])
+
+moment_tensor = rtf2xyz(moment_tensor_rtf)  #make_moment_tensor(strike, dip, rake, M0, 0, 0))
+source = pp.PointSource(0,0,depth,moment_tensor,np.zeros([3,1]),0)
+stf = lambda w: stf_trapezoidal(w, 4.2, 8.4)
+
+receivers = pp.RegularlyDistributedReceivers(1,100,nr,0,360,nphi,depth=0)
+xx,yy = receivers.as_xy()
+
+
+seismogram_file = 'data.bin'
+compute_seismograms=False # Set to True to compute seismograms rather than load
+if compute_seismograms:
+    time, seismograms = pp.compute_seismograms(model,source,receivers,nt,delta_t,
+                                            source_time_function=stf,stencil_kwargs=stencil_args)
+    if seismogram_file is not None:
+        with open(seismogram_file,'wb') as fp:
+            np.save(fp, time)
+            np.save(fp, seismograms)
+else:
+    print("Attempting to load seismograms from file:")
+    try:
+        with open(seismogram_file,'rb') as fp:
+            time = np.load(fp)
+            seismograms = np.load(fp)
+    except:
+        print("Failed. Consider re-running with `compute_seismograms=True`")
+        raise
+
+
+# Construct regular grid for interpolation of functions
+xgrid, ygrid = np.meshgrid(np.linspace(-box_size,box_size,ngrid),
+                                np.linspace(-box_size,box_size,ngrid))
+
+# Differentiate seismograms to get accelerograms and calculate overall amplitude -
+# used to plot wavefronts
+acc = np.linalg.norm(np.diff(seismograms,n=2),axis=2)
+maxacc = acc.max()
+
[…]
