+++ examples/idaho_wavefield_movie.py
@@ -1,145 +1,0 @@
-import numpy as np 
-import pyprop8 as pp
-from pyprop8.utils import rtf2xyz, make_moment_tensor,stf_trapezoidal
-import tqdm 
-from mayavi import mlab
-from scipy.interpolate import griddata
-
-print(pp.__file__)
-# This script generates an animated gif showing ground motion and
-# line-of-sight static displacement (~ InSAR) for an event
-lo_res = False # Set to True to generate a quick-and-dirty plot for checking layout etc
-if lo_res:
-    nt=180
-    delta_t = 0.25
-    ngrid=50
-    ncontour=28
-    dpi=100
-    framestep=5
-    nr = 20
-    nphi=36
-    stencil_args = {'kmin':0,'kmax':2,'nk':1200}
-else:
-    nt = 900      # Number of time-steps in seismogram calculation
-    delta_t = 0.05 # Time interval in seismogram calculation
-    ngrid=250     # Number of grid points in Cartesian grid for interpolation
-    dpi=300       # resolution of output
-    ncontour=140  # Number of contours in static displacement plot
-    framestep=1   # Stride for animation
-    nr=100         # Number of radii for seismogram calculation
-    nphi=360       # Number of azimuths for seismogram calculation
-    # To ensure a high-quality resolution of the surface, increase the
-    # spatial (k-space) sampling
-    stencil_args = {'kmin':0,'kmax':5,'nk':12000}
-box_size = 70
-# Earth model to use
-
-# These values are Based on Crust1.0 for Idaho
-model = pp.LayeredStructureModel([( 16.5, 6.1, 3.6 , 2.7),
-                                 ( 14.6, 6.3, 3.7, 2.8),
-                                 (  6.4, 7.0, 4.0, 3.0),
-                                 (np.inf,7.9, 4.4, 3.3)])
-
-# Earthquake parameters
-# GCMT catalog values
-depth = 13.8
-moment_tensor_rtf = 1e6*np.array([[-2.320, 1.120, 1.150],
-                                  [ 1.120, 1.780,-5.970],
-                                  [ 1.150,-5.970, 0.535]])
-
-moment_tensor = rtf2xyz(moment_tensor_rtf)  #make_moment_tensor(strike, dip, rake, M0, 0, 0))
-source = pp.PointSource(0,0,depth,moment_tensor,np.zeros([3,1]),0)
-stf = lambda w: stf_trapezoidal(w, 4.2, 8.4)
-
-receivers = pp.RegularlyDistributedReceivers(1,100,nr,0,360,nphi,depth=0)
-xx,yy = receivers.as_xy()
-
-if False:
-    time, seismograms = pp.compute_seismograms(model,source,receivers,nt,delta_t,source_time_function=stf,stencil_kwargs=stencil_args)
-    with open('data.bin','wb') as fp:
-        np.save(fp, time)
-        np.save(fp, seismograms)
-else:
-    with open('data.bin','rb') as fp:
-        time = np.load(fp)
-        seismograms = np.load(fp)
-
-
-xgrid, ygrid = np.meshgrid(np.linspace(-box_size,box_size,ngrid),np.linspace(-box_size,box_size,ngrid))
-insar = np.zeros([ngrid,ngrid,3])
[…]
