+++ examples/idaho_wavefield_movie/idaho_wavefield_movie.py
@@ -110,60 +110,65 @@ boxcolor=(0,0,0) # Color for box corner annotation
 ticksize=5 # Size of box corners
 
 # mayavi is supposed to do animations but I can't make it work properly. #ust generate each frame 
-# separately and then stitch them together at the end...
+# separately and then stitch them together at the end.
+#
+# The plotting code could doubtless be made cleaner and nicer...
 #
 # First a loop for plotting wavefield evolution over time
-for i in tqdm.tqdm(range(nt)):
-    # Clear the image and start again
-    mlab.clf()
+do_wavefield=True # Useful to be able to turn this off and skip straight to second loop
+if do_wavefield:
+    for i in tqdm.tqdm(range(nt)):
+        # Clear the image and start again
+        mlab.clf()
 
-    # Compute data for the image of wavefronts -> magnitude of acceleration at this instant
-    accstep = np.zeros([ngrid,ngrid])
-    if i==0 or i==nt-1:
-        pass # We have lost two time steps due to the numerical differentiation
-    else:
-        # Interpolate the acceleration data.
-        # Note that if timestep * wavespeed < distance between sampling points in xx/yy, aliasing will happen
-        accstep = griddata((xx.flatten(),yy.flatten()),acc[:,:,i-1].flatten(),(xgrid.flatten(),ygrid.flatten())).reshape((ngrid,ngrid))
-    
-    # Plot the main surface-displacement image
-    surf = mlab.mesh(xx+scale_horiz*seismograms[:,:,0,i],
-                     yy+scale_horiz*seismograms[:,:,1,i],
-                     scale_vert*seismograms[:,:,2,i],
-                     vmin=-cmax,vmax=cmax,
-                     colormap='RdYlBu',reset_zoom=False)
+        # Compute data for the image of wavefronts -> magnitude of acceleration at this instant
+        accstep = np.zeros([ngrid,ngrid])
+        if i==0 or i==nt-1:
+            pass # We have lost two time steps due to the numerical differentiation
+        else:
+            # Interpolate the acceleration data.
+            # Note that if timestep * wavespeed < distance between sampling points in xx/yy, aliasing will happen
+            accstep = griddata((xx.flatten(),yy.flatten()),acc[:,:,i-1].flatten(),(xgrid.flatten(),ygrid.flatten())).reshape((ngrid,ngrid))
+        
+        # Plot the main surface-displacement image
+        surf = mlab.mesh(xx+scale_horiz*seismograms[:,:,0,i],
+                        yy+scale_horiz*seismograms[:,:,1,i],
+                        scale_vert*seismograms[:,:,2,i],
+                        vmin=-cmax,vmax=cmax,
+                        colormap='seismic',reset_zoom=False)
 
-    # Plot corners of square region used for lower plot
-    mlab.plot3d([box_size-ticksize,box_size,box_size],[-box_size,-box_size,-box_size+ticksize],[hover,hover,hover],tube_radius=0.2,color=boxcolor)
-    mlab.plot3d([-box_size+ticksize,-box_size,-box_size],[-box_size,-box_size,-box_size+ticksize],[hover,hover,hover],tube_radius=0.2,color=boxcolor)
-    mlab.plot3d([box_size-ticksize,box_size,box_size],[box_size,box_size,box_size-ticksize],[hover,hover,hover],tube_radius=0.2,color=boxcolor)
[…]
