### 2022-08-10 valentineap/pyprop8#10 @valentineap
I've done various reorganisations of the code base. The main goal was to separate out 'user-facing' functions from the 'backend'. I've also tried to ensure that everything has some basic documentation, highlighting links to equations in the papers. A mathematica notebook (& pdf version) in `/notes/` explains where some pieces come from that are expressed in a different form from the paper.

I have thought about providing plotting scripts. However, these would really be very basic (`plt.plot(time, seismogram[i,0,:])`). Given the goal for pyprop8 to be minimal in scope I think it better not to add anything here. There are a number of scripts/notebooks in the examples directory that demonstrate how to plot things, should a user seek some guidance.

 No doubt more could be done to address the various comments here, but I hope that the changes are an improvement and sufficient to help someone make sense of the package.
