+++ paper.bib
@@ -119,3 +119,13 @@
   pages   = {261--272},
   doi     = {10.1038/s41592-019-0686-2},
 }
+
+@ARTICLE{Sambridge2022,
+	author = {Sambridge, M. and Jackson, A. and Valentine, A.P.},
+	title = {Geophysical inversion and Optimal Transport},
+	year = {2022},
+	journal = {Geophysical Journal International},
+	volume = {231},
+	pages = {172--198},
+	doi = {10.1093/gji/ggac151},
+}
