+++ paper.md
@@ -41,9 +41,9 @@ A number of assumptions or approximations are inherent to the formulation of `py
 These factors should be borne in mind in any case where outputs from `pyprop8` are to be compared to observational data.
 
 # Applications
-The computational approach described by @OToole2011 (but not the `pyprop8` implementation) has underpinned a number of studies---particularly those focussed on earthquake early warning using continuous GPS data [@OToole2012; @OToole2013; @Kaufl2014]. It has also been exploited for monitoring of microseismicity, e.g. during hydraulic fracturing [@OToole2013a]. We anticipate that the release of `pyprop8` will enable further applications of this kind.
+The computational approach described by @OToole2011 has underpinned a number of studies---particularly those focussed on earthquake early warning using continuous GPS data [@OToole2012; @OToole2013; @Kaufl2014]. It has also been exploited for monitoring of microseismicity, e.g. during hydraulic fracturing [@OToole2013a]. (For the avoidance of doubt: these works were based on a Fortran77 implementation of @OToole2011, rather than  `pyprop8`.) We anticipate that the release of `pyprop8` will enable further applications of this kind.
 
-Much of our work focusses on the development of new strategies for solving inverse problems. In this context, `pyprop8` provides a valuable resource for implementing test problems. In particular, it offers scalable complexity across both model and data spaces. It can support examples that are linear, weakly non-linear, or highly non-linear, and datasets may range from isolated slip-vectors through to dense arrays of time-series. This allows a structured approach to algorithm development, progressing from simpler to more complex test problems [@Kaufl2014; @Kaufl2015; @Kaufl2016a].
+Much of our work focusses on the development of new strategies for solving inverse problems. In this context, `pyprop8` provides a valuable resource for implementing test problems. In particular, it offers scalable complexity across both model and data spaces. It can support examples that are linear, weakly non-linear, or highly non-linear, and datasets may range from isolated slip-vectors through to dense arrays of time-series. This allows a structured approach to algorithm development, progressing from simpler to more complex test problems [@Kaufl2014; @Kaufl2015; @Kaufl2016a]. As such, `pyprop8` underpins many of the examples presented in @Sambridge2022, which develops a novel optimal-transport based misfit function for waveform inversion.
 
 # Acknowledgements
 The theory described in @OToole2011 was originally implemented in Fortran 77 by John Woodhouse and Tom O'Toole. We have benefited from access to a copy of this code (`prop8`), which was distributed informally but never publicly released. This has informed certain aspects of our own implementation, and enabled us to validate our version against theirs. Our work has benefited from financial support from the Australian Research Council under grants DE180100040 and DP200100053, and from the CSIRO Future Science Platform in Deep Earth Imaging.
