### 2022-08-10 f593289
Added figure to paper
