### 2022-08-10 1fa9b75
Rescaled figure
