+++ paper.md
@@ -29,7 +29,7 @@ Many different tools and packages exist which may be used to simulate seismic wa
 To address this, `pyprop8` is written entirely in Python, and the core seismogram-calculation routines do not have any dependencies except the mainstream libraries `numpy` and `scipy`. We envisage that this package will serve a diverse range of users. For those studying or teaching seismology, `pyprop8` provides a simulation tool with a reasonable level of physical realism and flexibility, which can underpin a wide range of demonstrations and practical exercises. Similarly, for those studying or teaching inverse theory, `pyprop8` may be used as an exemplar forward problem with realistic features and scalable complexity. In this context, the ability to compute partial derivatives with respect to source parameters is valuable. The package may also prove beneficial to the research community, particularly as a source of test problems for use in the development of inversion algorithms.
 
 # Implementation details
-![Seismic observables computed using `pyprop8`. Surface displacement (top) and wrapped line-of-sight displacement simulated for the Stanley, Idaho earthquake of 31 March 2020.\label{fig:example}](./paper_figure.png)
+![Seismic observables computed using `pyprop8`. Surface displacement (top) and wrapped line-of-sight displacement simulated for the Stanley, Idaho earthquake of 31 March 2020.\label{fig:example}](./paper_figure.png){ width=70% }
 
 The theoretical basis for `pyprop8` is described in detail in @OToole2011, with the calculation of derivatives set out in an appendix to @OToole2012. These papers build on earlier work by @Woodhouse1980a. The present package provides a full implementation of this published theory. The algorithm is based on a Thompson-Haskell propagator matrix method, using a minor-vector formalism to ensure numerical stability. A novel feature of the approach is that it is designed to remain stable at zero frequency, allowing the static offset (i.e., permanent seismic deformation) to be captured within simulations (see \autoref{fig:example}.). As a result, a wide range of seismic observables can be simulated within a common framework, including conventional seismograms, GPS records, InSAR images and direct field observations of slip.
 
