### 2022-08-10 valentineap/pyprop8#12 @valentineap
I've rephrased this bit of the paper. I believe `pyprop8` implements everything from the O'Toole et al. papers. I've also added a nice figure!
