### 2022-08-10 #4217 @valentineap
Hi @leouieda, I think I've finally dealt with all the comments and suggestions. Please see issues tagged [joss-paper-reviewer-comment](https://github.com/valentineap/pyprop8/issues?q=label%3Ajoss-paper-reviewer-comment+is%3Aclosed) on the issue tracker for my responses.

Note that I have not yet updated PyPI package, in case any further changes are needed.

Let me know if you need anything else from me!
