### 2022-08-10 #4217 @hemmelig
I've had a look through the changes made in response to my review and am happy to say everything has been addressed. The new tests and the Idaho video example are great! I installed a fresh copy and ran the examples and everything worked just fine for me (Ubuntu 20.04, Python 3.8).

I realise the internal documentation requires a lot of time and effort, but it has definitely made it much clearer how the different functions relate to the ideas/maths set out in the papers (particularly with the additional notes)!

Conor
