### 2022-08-10 #4217 @hfmark
I'll echo Conor - @valentineap, it looks like you've addressed all my comments and then some. The updated paper text looks great, and the code all seems to work on my system (Ubuntu 18.04, Python 3.9) including the excellent new test suite. My checklist is all set!
