### 2022-08-11 7d36b9c
Fixed error in annotation of movie/stills
