### 2022-08-23 659efe7
Update setup.cfg
Increment version number
