### 2022-08-24 #4217 @leouieda
@valentineap I'm very happy with the state of the paper/software and the reviewers are satisfied as well. It's my pleasure to move on to acceptance of your paper 🎉

There just a few of things that you'll need to do:

1. Double check author names and affiliations (including ORCIDs)
1. Make a release of the software with the latest changes from the review (including any editorial fixes). This is the version that will be used in the JOSS paper. **Post the version number here** once it's available.
1. Archive the release on [Zenodo](https://zenodo.org) or equivalent repository and **post the DOI here**.
1. Make sure that the **title and author list** (including ORCIDs) in the archive above **match those in the JOSS paper**.

Once those are done, we can move on to publication! Let me know if you need any help with these or have any questions.
