### 2022-08-24 ca19c1a
Update setup.cfg
