+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = pyprop8
-version = 1.1.1
+version = 1.1.2
 author = Andrew Valentine
 author_email = andrew.valentine@durham.ac.uk
 description = A lightweight package for synthetic seismograms
