### 2022-08-24 #4217 @valentineap
Excellent! Thanks again @leouieda, @hfmark and @hemmelig - I really appreciate your input into this.

1. I confirm that the authors and ORCIDs in the manuscript are correct
2. After a bit of PyPI and Zenodo silliness I think we are at [version 1.1.2](https://github.com/valentineap/pyprop8/releases/tag/v1.1.2) which is now [available with `pip`](https://pypi.org/project/pyprop8/)
3. [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.7019949.svg)](https://doi.org/10.5281/zenodo.7019949)
4. After some more silliness... I think they match!

So -- I hope that's all done, but let me know if there are any problems.
